+++ nvim/lua/custom/plugins/bufferline.lua
@@ -1,0 +1,13 @@
+-- Shows a bufferline at top of file
+return {
+  'akinsho/bufferline.nvim',
+  version = '*',
+  dependencies = 'nvim-tree/nvim-web-devicons',
+  config = function()
+    require('bufferline').setup {
+      options = {
+        numbers = 'buffer_id',
+      },
+    }
+  end,
+}