+++ nvim/lua/custom/plugins/nvim-genghis.lua
@@ -1,0 +1,1 @@
+return { 'chrisgrieser/nvim-genghis' }